# specs/openapi.yaml
openapi: 3.0.3
info:
  title: Приложение для федерации тяжелой атлетики г. Москвы (ФТАМ)
  description: Приложение для спортcменов, тренеров, организаторов и представителей ФТАМ - автоматизация деятельности ФТАМ.
  version: 1.0.11

# servers:
#   - url: https://wlmoscow
...
paths:
  /sportsmen/{id}/results:
    get:
      tags:
        - sportsmen
      summary: Получить все результаты спортсмена
      operationId: getResultsBySmId
      parameters:
        - name: id
          in: path
          description: ID спортсмена
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllSmResults'          
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
...
  /tcamps:
    get:
      tags:
        - tcamps
      summary: Получить все сборы
      operationId: getTCamps
      parameters:
        - name: date
          in: query
          description: Дата начала сборов
          required: false
          schema:
            type: string
            format: datetime(utc)
        - name: city
          in: query
          description: Город
          required: false
          schema:
            example: "Москва"
            type: string
        - name: sort
          in: query
          description: Сортировка. Список имен атрибутов, разделенных запятыми, для сортировки. Если перед именем атрибута стоит знак минус, то сортировка производится по убыванию, в противном случае — по возрастанию.
          required: false
          schema:
            example: date,-city
            type: string
        - name: page
          in: query
          description: Пагинация - номер страницы.
          required: false
          schema:
            example: 3
            type: integer
        - name: batch
          in: query
          description: Пагинация - количество элементов на странице
          required: false
          schema:
            example: 10
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllTCamps'
        '500':
          description: Internal Server Error
    post:
      tags:
        - tcamps
      summary: Организация сборов
      operationId: createTCamp
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTCamp'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TCamp' 
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /tcamps/{id}:
    delete:
      tags:
        - tcamps
      summary: Удалить сборы
      operationId: deleteTCamp
      parameters:
        - name: id
          in: path
          description: ID сборов
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
            description: Successful operation
        '401':
          description: Unauthorized
        '404':
            description: Not found
        '500':
            description: Internal Server Error
...
  /sportsmen:
    get:
      tags:
        - sportsmen
      summary: Получить всех спортсменов
      description: Возможно в параметрах нужно полное ФИО
      operationId: getSportsmen
      parameters:
        - name: surname
          in: query
          description: Фамилия
          required: false
          schema:
            type: string
            example: "Иванов"
        - name: name
          in: query
          description: Имя
          required: false
          schema:
            type: string
            example: "Иван"
        - name: patronimyc
          in: query
          description: Отчество
          required: false
          schema:
            type: string
            example: "Иванович"
        - name: sort
          in: query
          description: Сортировка. Список имен атрибутов, разделенных запятыми, для сортировки. Если перед именем атрибута стоит знак минус, то сортировка производится по убыванию, в противном случае — по возрастанию.
          required: false
          schema:
            example: -name
            type: string
        - name: page
          in: query
          description: Пагинация - номер страницы.
          required: false
          schema:
            example: 3
            type: integer
        - name: batch
          in: query
          description: Пагинация - количество элементов на странице
          required: false
          schema:
            example: 10
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllSportsmen'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
...
  /sportsmen/{id}:
    get:
      tags:
        - sportsmen
      summary: Получить спортсмена по ID
      operationId: getSportsmanById
      parameters:
        - name: id
          in: path
          description: ID спортсмена
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SportsmanExt'
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: Sportsman not found
        '500':
          description: Internal Server Error
    patch:
      tags:
        - sportsmen
      summary: Обновление секретарем информации о спортсмене
      description: Обновление информации о сертификате и допуске на соревнования
      operationId: UpdateSportsman
      parameters:
          - name: id
            in: path
            description: ID спортсмена
            required: true
            schema:
              type: string
              format: uuid
      requestBody:
        description: Обновление информации о существующем спортсмене
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSportsman'
        required: false
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SportsmanExt'
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: Sportsman not found
        '500':
          description: Internal Server Error
    delete:
      tags:
        - sportsmen
      summary: Удалить спортсмена
      operationId: deleteSportsman
      parameters:
        - name: id
          in: path
          description: ID спортсмена
          required: true
          schema:
              type: string
              format: uuid
      responses:
        '204':
            description: Successful operation
        '401':
          description: Unauthorized
        '404':
            description: Sportsman not found
        '500':
            description: Internal Server Error
...
  /sportsmen/{id}/coach:
    post:
      tags:
        - sportsmen
      summary: Запись спортсмена к тренеру
      parameters:
        - name: id
          in: path
          description: ID спортсмена
          required: true
          schema:
              type: string
              format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CId'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmCoach'
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Internal Server Error
...
  /tcamps/{id}/sportsman:
    post:
      tags:
        - tcamps
      summary: Запись спортсмена на сборы
      parameters:
        - name: id
          in: path
          description: ID сборов
          required: true
          schema:
              type: string
              format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTCampApplication'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TCampApplication'
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Internal Server Error
...
components:
  schemas:
    TCamp:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        city:
          type: string
          example: Москва
        address:
          type: string
          example: Московская 1
        beg_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
...
    Sportsman:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        surname:
          type: string
          example: Иванов
        name:
          type: string
          example: Иван
        patronymic:
          type: string
          example: Иванович
        birthday:
          type: string
          format: date-time
        moscow_team:
          type: boolean
          example: true
        sports_category:
          type: string
          example: КМС
        gender:
          type: boolean
          example: false
    SportsmanExt:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        surname:
          type: string
          example: Иванов
        name:
          type: string
          example: Иван
        patronymic:
          type: string
          example: Иванович
        birthday:
          type: string
          format: date-time
        moscow_team:
          type: boolean
          example: true
        sports_category:
          type: string
          example: КМС
        gender:
          type: boolean
          example: false
        adoping_validity:
          type: string
          format: date-time
        access_validity:
          type: string
          format: date-time
        access_institution:
          type: string
          example: 1-й диспансер
...
    TCampApplication:
      type: object
      properties:
        tcamp_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440001
        sm_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
    CreateTCampApplication:
      type: object
      properties:
        sm_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
...
    CreateTCamp:
      type: object
      properties:
        city:
          type: string
          example: Москва
        address:
          type: string
          example: Московская 1
        beg_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
    UpdateSportsman:
      type: object
      properties:
        moscow_team:
          type: boolean
          example: true
        sports_category:
          type: string
          example: КМС
        access_institution:
          type: string
          example: 1-й диспансер
        access_validity:
          type: string
          format: date-time
        antidoping_validity:
          type: string
          format: date-time
    AllTCamps:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/TCamp'
...
    AllSportsmen:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/Sportsman'
...
    SmResult:
      type: object
      properties:
        comp_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440001
        comp_name:
          type: string
          example: Кубок Москвы
        sm_fullname:
          type: string
          example: Иванов Иван Иванович
        beg_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        city:
          type: string
          example: Москва
        weight_category:
          type: string
          example: 109+
        snatch:
          type: integer
          example: 120
        clean_and_jerk:
          type: integer
          example: 200
        place:
          type: integer
          example: 2
    AllSmResults:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/SmResult'
...
    CId:
      type: object
      properties:
        c_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440001
    SmCoach:
      type: object
      properties:
        sm_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        c_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440001
...